+++ data-prepper-plugins/kinesis-source/src/test/java/org/opensearch/dataprepper/plugins/kinesis/source/KinesisMultiStreamTrackerTest.java
@@ -10,7 +10,6 @@
 
 package org.opensearch.dataprepper.plugins.kinesis.source;
 
-import com.linecorp.armeria.client.retry.Backoff;
 import com.google.common.collect.ImmutableList;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -19,11 +18,6 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.opensearch.dataprepper.plugins.kinesis.source.configuration.KinesisSourceConfig;
 import org.opensearch.dataprepper.plugins.kinesis.source.configuration.KinesisStreamConfig;
-import org.opensearch.dataprepper.plugins.kinesis.source.exceptions.KinesisRetriesExhaustedException;
-import software.amazon.awssdk.services.kinesis.KinesisAsyncClient;
-import software.amazon.awssdk.services.kinesis.model.DescribeStreamSummaryRequest;
-import software.amazon.awssdk.services.kinesis.model.DescribeStreamSummaryResponse;
-import software.amazon.awssdk.services.kinesis.model.StreamDescriptionSummary;
 import software.amazon.kinesis.common.InitialPositionInStream;
 import software.amazon.kinesis.common.InitialPositionInStreamExtended;
 import software.amazon.kinesis.common.StreamConfig;
@@ -36,25 +30,18 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.CompletableFuture;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyInt;
-import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 public class KinesisMultiStreamTrackerTest {
     private static final String APPLICATION_NAME = "multi-stream-application";
     private static final String awsAccountId = "1234";
-    private static final String streamArnFormat = "arn:aws:kinesis:us-east-1:%s:stream/%s";
     private static final Instant streamCreationTime = Instant.now();
     private static final List<String> STREAMS_LIST = ImmutableList.of("stream-1", "stream-2", "stream-3");
 
-    @Mock
-    private KinesisAsyncClient kinesisClient;
     private List<StreamConfig> streamConfigList;
 
     private Map<String, KinesisStreamConfig> streamConfigMap;
@@ -63,7 +50,7 @@ public class KinesisMultiStreamTrackerTest {
     KinesisSourceConfig kinesisSourceConfig;
 
     @Mock
-    private Backoff backoff;
+    private KinesisStreamBackoffStrategy kinesisStreamBackoffStrategy;
 
     @BeforeEach
     public void setUp() {
@@ -74,25 +61,8 @@ public class KinesisMultiStreamTrackerTest {
             KinesisStreamConfig kinesisStreamConfig = mock(KinesisStreamConfig.class);
             when(kinesisStreamConfig.getName()).thenReturn(stream);
             when(kinesisStreamConfig.getInitialPosition()).thenReturn(InitialPositionInStream.LATEST);
-
-            StreamDescriptionSummary streamDescriptionSummary = StreamDescriptionSummary.builder()
-                    .streamARN(String.format(streamArnFormat, awsAccountId, stream))
-                    .streamCreationTimestamp(streamCreationTime)
-                    .streamName(stream)
-                    .build();
-
-            DescribeStreamSummaryRequest describeStreamSummaryRequest = DescribeStreamSummaryRequest.builder()
-                    .streamName(stream)
-                    .build();
-
-            DescribeStreamSummaryResponse describeStreamSummaryResponse = DescribeStreamSummaryResponse.builder()
-                    .streamDescriptionSummary(streamDescriptionSummary)
-                    .build();
-
-            when(kinesisClient.describeStreamSummary(describeStreamSummaryRequest))
-                    .thenReturn(CompletableFuture.completedFuture(describeStreamSummaryResponse));
+            when(kinesisStreamBackoffStrategy.getStreamIdentifier(stream)).thenReturn(getStreamIdentifier(stream));
             kinesisStreamConfigs.add(kinesisStreamConfig);
-
             streamConfigMap.put(stream, kinesisStreamConfig);
         });
 
@@ -100,7 +70,7 @@ public class KinesisMultiStreamTrackerTest {
     }
 
     private KinesisMultiStreamTracker createObjectUnderTest() {
-        return new KinesisMultiStreamTracker(kinesisClient, kinesisSourceConfig, APPLICATION_NAME, backoff);
+        return new KinesisMultiStreamTracker(kinesisSourceConfig, APPLICATION_NAME, kinesisStreamBackoffStrategy);
     }
 
     @Test
@@ -129,13 +99,8 @@ public class KinesisMultiStreamTrackerTest {
             when(kinesisStreamConfig.getName()).thenReturn(stream);
             when(kinesisStreamConfig.getInitialPosition()).thenReturn(InitialPositionInStream.LATEST);
 
-            DescribeStreamSummaryRequest describeStreamSummaryRequest = DescribeStreamSummaryRequest.builder()
-                    .streamName(stream)
-                    .build();
-
-            when(kinesisClient.describeStreamSummary(describeStreamSummaryRequest)).thenThrow(new RuntimeException());
+            when(kinesisStreamBackoffStrategy.getStreamIdentifier(stream)).thenThrow(new RuntimeException());
             kinesisStreamConfigs.add(kinesisStreamConfig);
-
             streamConfigMap.put(stream, kinesisStreamConfig);
         });
 
@@ -155,35 +120,8 @@ public class KinesisMultiStreamTrackerTest {
             when(kinesisStreamConfig.getName()).thenReturn(stream);
             when(kinesisStreamConfig.getInitialPosition()).thenReturn(InitialPositionInStream.LATEST);
 
-            StreamDescriptionSummary streamDescriptionSummary = StreamDescriptionSummary.builder()
-                    .streamARN(String.format(streamArnFormat, awsAccountId, stream))
-                    .streamCreationTimestamp(streamCreationTime)
-                    .streamName(stream)
-                    .build();
-
-            DescribeStreamSummaryRequest describeStreamSummaryRequest = DescribeStreamSummaryRequest.builder()
-                    .streamName(stream)
-                    .build();
-
-            DescribeStreamSummaryResponse describeStreamSummaryResponse = DescribeStreamSummaryResponse.builder()
-                    .streamDescriptionSummary(streamDescriptionSummary)
-                    .build();
-            final CompletableFuture<DescribeStreamSummaryResponse> successFuture = CompletableFuture.completedFuture(describeStreamSummaryResponse);
-
-            kinesisStreamConfigs.add(kinesisStreamConfig);
-            streamConfigMap.put(stream, kinesisStreamConfig);
-
-            final CompletableFuture<DescribeStreamSummaryResponse> failedFuture1 = new CompletableFuture<>();
-            failedFuture1.completeExceptionally(mock(Throwable.class));
-            final CompletableFuture<DescribeStreamSummaryResponse> failedFuture2 = new CompletableFuture<>();
-            failedFuture2.completeExceptionally(mock(Throwable.class));
-
-            given(kinesisClient.describeStreamSummary(describeStreamSummaryRequest))
-                    .willReturn(failedFuture1)
-                    .willReturn(failedFuture2)
-                    .willReturn(successFuture);
-
-            when(backoff.nextDelayMillis(anyInt())).thenReturn(100L);
+            StreamIdentifier streamIdentifier = getStreamIdentifier(stream);
+            when(kinesisStreamBackoffStrategy.getStreamIdentifier(stream)).thenReturn(streamIdentifier);
         });
 
         when(kinesisSourceConfig.getStreams()).thenReturn(kinesisStreamConfigs);
@@ -199,41 +137,6 @@ public class KinesisMultiStreamTrackerTest {
             assertEquals(streamIdentifier, getStreamIdentifier(stream));
             assertEquals(initialPositionInStreamExtended, InitialPositionInStreamExtended.newInitialPosition(streamConfigMap.get(stream).getInitialPosition()));
         }
-    }
-
-    @Test
-    public void testStreamConfigWithRetriesThrowsException() {
-        MockitoAnnotations.openMocks(this);
-        List<KinesisStreamConfig> kinesisStreamConfigs = new ArrayList<>();
-        streamConfigMap = new HashMap<>();
-        STREAMS_LIST.forEach(stream -> {
-            KinesisStreamConfig kinesisStreamConfig = mock(KinesisStreamConfig.class);
-            when(kinesisStreamConfig.getName()).thenReturn(stream);
-            when(kinesisStreamConfig.getInitialPosition()).thenReturn(InitialPositionInStream.LATEST);
-
-            DescribeStreamSummaryRequest describeStreamSummaryRequest = DescribeStreamSummaryRequest.builder()
-                    .streamName(stream)
-                    .build();
-
-            when(backoff.nextDelayMillis(anyInt())).thenReturn((long) -1);
-
-            final CompletableFuture<DescribeStreamSummaryResponse> failedFuture1 = new CompletableFuture<>();
-            failedFuture1.completeExceptionally(mock(Throwable.class));
-            final CompletableFuture<DescribeStreamSummaryResponse> failedFuture2 = new CompletableFuture<>();
-            failedFuture2.completeExceptionally(mock(Throwable.class));
-            final CompletableFuture<DescribeStreamSummaryResponse> failedFuture3 = new CompletableFuture<>();
-            failedFuture3.completeExceptionally(mock(Throwable.class));
-
-            when(kinesisClient.describeStreamSummary(describeStreamSummaryRequest))
-                    .thenReturn(failedFuture1)
-                    .thenReturn(failedFuture2)
-                    .thenReturn(failedFuture3);
-
-            kinesisStreamConfigs.add(kinesisStreamConfig);
-            streamConfigMap.put(stream, kinesisStreamConfig);
-        });
-        when(kinesisSourceConfig.getStreams()).thenReturn(kinesisStreamConfigs);
-        assertThrows(KinesisRetriesExhaustedException.class, () -> createObjectUnderTest().streamConfigList());
     }
 
     @Test
